CodeCov more reliable then Coveralls
Coverall seems to randomly fail to analyze the code and repros 0% code coverage and a failure on the commit.  I've also seen the badge in the readme.md not get updated and show 0% code coverage even when there are newer builds that have 100% code coverage.  CodeCov seems to be much more reliable and consistent, so removing Coveralls.